Always require verification if user has a device
Adds a check for user devices in VerifyUserPermissionsMiddleware.
If the user has a device, always require them to verify, even if they
don't have 2FA enabled (they are not in a group with the permission).

Without this check, if a user leaves a group which has 2FA enabled, they
would no longer have to verify using their device.